b787: test hungarian and uno algorithms, lazy load networkx
