+++ qcelemental/util/__init__.py
@@ -1,5 +1,7 @@
 from .np_blockwise import blockwise_expand, blockwise_contract
 from .np_rand3drot import random_rotation_matrix
+from .scipy_hungarian import linear_sum_assignment
+from .gph_uno_bipartite import uno
 #from .mpl import plot_coord
 from .misc import (distance_matrix, update_with_error, standardize_efp_angles_units, filter_comments, unnp,
                    compute_distance, compute_angle, compute_dihedral, measure_coordinates)
